+++ plugin/src/main/java/com/stratio/cassandra/lucene/IndexServiceSkinny.java
@@ -15,8 +15,8 @@
  */
 package com.stratio.cassandra.lucene;
 
-import com.stratio.cassandra.lucene.core.column.Columns;
-import com.stratio.cassandra.lucene.core.column.ColumnsMapper;
+import com.stratio.cassandra.lucene.column.Columns;
+import com.stratio.cassandra.lucene.column.ColumnsMapper;
 import com.stratio.cassandra.lucene.index.DocumentIterator;
 import com.stratio.cassandra.lucene.key.PartitionMapper;
 import org.apache.cassandra.db.*;
