+++ plugin/src/main/java/com/stratio/cassandra/lucene/schema/Schema.java
@@ -17,8 +17,8 @@ package com.stratio.cassandra.lucene.schema;
 
 import com.google.common.base.MoreObjects;
 import com.stratio.cassandra.lucene.IndexException;
-import com.stratio.cassandra.lucene.core.column.Column;
-import com.stratio.cassandra.lucene.core.column.Columns;
+import com.stratio.cassandra.lucene.column.Column;
+import com.stratio.cassandra.lucene.column.Columns;
 import com.stratio.cassandra.lucene.schema.mapping.Mapper;
 import com.stratio.cassandra.lucene.search.Search;
 import org.apache.cassandra.config.ColumnDefinition;
