+++ plugin/src/main/java/com/stratio/cassandra/lucene/schema/mapping/BitemporalMapper.java
@@ -17,8 +17,8 @@ package com.stratio.cassandra.lucene.schema.mapping;
 
 import com.google.common.base.MoreObjects;
 import com.stratio.cassandra.lucene.IndexException;
-import com.stratio.cassandra.lucene.core.column.Column;
-import com.stratio.cassandra.lucene.core.column.Columns;
+import com.stratio.cassandra.lucene.column.Column;
+import com.stratio.cassandra.lucene.column.Columns;
 import com.stratio.cassandra.lucene.util.DateParser;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.lucene.document.LongField;
