+++ plugin/src/test/java/com/stratio/cassandra/lucene/schema/mapping/DateRangeMapperTest.java
@@ -16,7 +16,7 @@
 package com.stratio.cassandra.lucene.schema.mapping;
 
 import com.stratio.cassandra.lucene.IndexException;
-import com.stratio.cassandra.lucene.core.column.Columns;
+import com.stratio.cassandra.lucene.column.Columns;
 import com.stratio.cassandra.lucene.schema.mapping.builder.DateRangeMapperBuilder;
 import org.apache.cassandra.utils.UUIDGen;
 import org.apache.lucene.document.Field;
